# docs/api-docs.json
{
"openapi": "3.0.1",
"info": {
"title": "TradingSim API",
"description": "交易模拟系统后端API文档",
"contact": {
"name": "TradingSim Team",
"url": "https://github.com/tradingsim/tradingsim",
"email": "support@tradingsim.com"
},
"license": {
"name": "MIT License",
"url": "https://opensource.org/licenses/MIT"
},
"version": "1.0.0"
},
"servers": [
{
"url": "http://localhost:8080",
"description": "开发环境"
},
{
"url": "https://api.tradingsim.com",
"description": "生产环境"
}
],
"security": [
{
"Bearer Authentication": []
}
],
...
"paths": {
...
"/data-import/latest/{stockCode}": {
"post": {
"tags": [
"数据导入"
],
"summary": "导入最新数据",
"description": "导入指定股票的最新数据",
"operationId": "importLatestData",
"parameters": [
{
"name": "stockCode",
"in": "path",
"description": "股票代码",
"required": true,
"schema": {
"type": "string"
},
"example": "AAPL"
}
],
"responses": {
"200": {
"description": "OK",
"content": {
"*/*": {
"schema": {
"type": "object",
"additionalProperties": {
"type": "object"
}
}
}
}
}
}
}
},
"/data-import/latest/batch": {
"post": {
"tags": [
"数据导入"
],
"summary": "批量导入最新数据",
"description": "批量导入多个股票的最新数据",
"operationId": "batchImportLatestData",
"parameters": [
{
"name": "stockCodes",
"in": "query",
"description": "股票代码列表",
"required": true,
"schema": {
"type": "array",
"items": {
"type": "string"
}
}
}
],
"responses": {
"200": {
"description": "OK",
"content": {
"*/*": {
"schema": {
"$ref": "#/components/schemas/ImportResult"
}
}
}
}
}
}
},
"/data-import/historical/{stockCode}": {
"post": {
"tags": [
"数据导入"
],
"summary": "导入历史数据",
"description": "导入指定股票在指定时间范围内的历史数据",
"operationId": "importHistoricalData",
"parameters": [
{
"name": "stockCode",
"in": "path",
"description": "股票代码",
"required": true,
"schema": {
"type": "string"
},
"example": "AAPL"
},
{
"name": "startTime",
"in": "query",
"description": "开始时间",
"required": true,
"schema": {
"type": "string",
"format": "date-time"
}
},
{
"name": "endTime",
"in": "query",
"description": "结束时间",
"required": true,
"schema": {
"type": "string",
"format": "date-time"
}
}
],
"responses": {
"200": {
"description": "导入成功",
"content": {
"*/*": {
"schema": {
"type": "object",
"additionalProperties": {
"type": "object"
}
}
}
}
},
"400": {
"description": "请求参数错误",
"content": {
"*/*": {
"schema": {
"type": "object",
"additionalProperties": {
"type": "object"
}
}
}
}
},
"500": {
"description": "导入失败",
"content": {
"*/*": {
"schema": {
"type": "object",
"additionalProperties": {
"type": "object"
}
}
}
}
}
}
}
},
"/data-import/historical/batch": {
"post": {
"tags": [
"数据导入"
],
"summary": "批量导入历史数据",
"description": "批量导入多个股票在指定时间范围内的历史数据",
"operationId": "batchImportHistoricalData",
"parameters": [
{
"name": "stockCodes",
"in": "query",
"description": "股票代码列表",
"required": true,
"schema": {
"type": "array",
"items": {
"type": "string"
}
}
},
{
"name": "startTime",
"in": "query",
"description": "开始时间",
"required": true,
"schema": {
"type": "string",
"format": "date-time"
}
},
{
"name": "endTime",
"in": "query",
"description": "结束时间",
"required": true,
"schema": {
"type": "string",
"format": "date-time"
}
}
],
"responses": {
"200": {
"description": "OK",
"content": {
"*/*": {
"schema": {
"$ref": "#/components/schemas/ImportResult"
}
}
}
}
}
}
},
...
"/data-import/integrity/{stockCode}": {
"get": {
"tags": [
"数据导入"
],
"summary": "检查数据完整性",
"description": "检查指定股票在指定时间范围内的数据完整性",
"operationId": "checkDataIntegrity",
"parameters": [
{
"name": "stockCode",
"in": "path",
"description": "股票代码",
"required": true,
"schema": {
"type": "string"
},
"example": "AAPL"
},
{
"name": "startTime",
"in": "query",
"description": "开始时间",
"required": true,
"schema": {
"type": "string",
"format": "date-time"
}
},
{
"name": "endTime",
"in": "query",
"description": "结束时间",
"required": true,
"schema": {
"type": "string",
"format": "date-time"
}
}
],
"responses": {
"200": {
"description": "OK",
"content": {
"*/*": {
"schema": {
"$ref": "#/components/schemas/DataIntegrityReport"
}
}
}
}
}
}
},
...
},
"components": {
"schemas": {
...
"ImportResult": {
"type": "object",
"properties": {
"totalStocks": {
"type": "integer",
"format": "int32"
},
"successfulStocks": {
"type": "integer",
"format": "int32"
},
"failedStocks": {
"type": "integer",
"format": "int32"
},
"totalRecords": {
"type": "integer",
"format": "int32"
},
"errors": {
"type": "array",
"items": {
"type": "string"
}
},
"successRate": {
"type": "number",
"format": "double"
}
}
},
...
"DataIntegrityReport": {
"type": "object",
"properties": {
"stockCode": {
"type": "string"
},
"startTime": {
"type": "string",
"format": "date-time"
},
"endTime": {
"type": "string",
"format": "date-time"
},
"expectedRecords": {
"type": "integer",
"format": "int64"
},
"actualRecords": {
"type": "integer",
"format": "int64"
},
"completeness": {
"type": "number",
"format": "double"
},
"missingPeriods": {
"type": "array",
"items": {
"type": "string"
}
},
"complete": {
"type": "boolean"
}
}
}
},
"securitySchemes": {
"Bearer Authentication": {
"type": "http",
"description": "请在请求头中添加 Authorization: Bearer {token}",
"scheme": "bearer",
"bearerFormat": "JWT"
}
}
}
}